feat: auth endpoint
Adds most of the code needed for the `auth` endpoint, which, given a
signed login token from a user, will issue a JWT access token signed by
the server authority.

This token can  be submitted to acess access-controlled endpoints.

An access token is only granted if:
* The login token is properly signed by `token.address`'s private key.
* The nonce is strictly increasing from the previous nonce (prevents
  replay attacks).
* The public key already exists (prevents spam attacks). To register an
  address on the DB, you'd have to send a payment first (hence, spam
  prevention).

This impl uses a default expiry preiod of 24hrs, but the ability to
configure this in the server would be simple to implement.

tests: auth endpoint