feat: acl middleware
Adds ACL middleware that offers fine-grain control for each route as to
which Roles are required in order to continue with the route.

To use it, specify the route using the `route!` macro:

```
route!(account => Get "/account/{address}" impl AccountManagement where requires [Role::ReadAll]);
```

Several tests are included that confirm that behaviour is as expected.