feat: accounts endpoint
support for
  /api/account
  /api/account/{address}

/api/account

allows users to query the status of their own account. The account is
derived from the Tari address contained in the access token metadata.

The returned JSON object corresponds to

pub struct UserAccount {
    pub id: i64,
    pub created_at: DateTime<Utc>,
    pub updated_at: DateTime<Utc>,
    pub total_received: MicroTari,
    pub total_pending: MicroTari,
    pub current_balance: MicroTari,
    pub total_orders: MicroTari,
}

/api/account/{address}

is an admin endpoint accessible by users with the ReadAll or SuperAdmin
role. Anyone else receives a 403 response when hitting the endpoint.

The returned object is the same as the above endpoint.

Testing

The following scenraios are added as tests:

* Unuathenticated user cannot access the `account` endpoint
* Standard user cannot access another user's account
* User with ReadAll role can access another account
* SuperAdmin role can access another account
* Standard user can access their own account