feat: track current orders
Adds a new field to the accounts table so that we can track the total
value of outstanding orders on the account.

current_orders is the debit-side equivalent of current_balance.

feat: add wallet auth tables

Adds hot wallet authorisation tables. Used for tracking payment
notifications from the hot wallet(s) and for preventing illegitimate
access to the payment notification endpoints.