feat: events subsystem
Adds an events subsystem (a simple pub/sub event emitter) to make it
easy to hook up external services in response to events (such as order
received, order paid, payment received etc).

The tests demonstrate how to set up the hooks. For example, the cucumber
tests set a `last_event` variable in tests that a `then` step can query
to make assertions about the emission of an event.