feat: update memo endpoint
Provides an /order_memo endpoint, which allows Admins to update the memo
field of an order.

Note that the HTTP method used for this endpoint is PATCH, rather than
POST.

The side effects of this call are:
* The memo is updated
* An OrderModifiedEvent is triggered.
* An audit log entry is added.

The memo is *not* checked for a valid signature and an order matching
cycle is not fired.

If a user has messed up the memo field, then we recommend cancelling the
order and asking the user to try again.

If this becomes cumbersome, and there's a clean flow for admins helping
provide a valid order signature, then we can modifiy this endpoint to do
so. Right now, it's not clear whether the UX would be any better than
re-doing the order.