feat: order reset
Provides an endpoint for admins to reset an order to the `New` state.

`reset_order` is a PATCH HTTP method.

This is useful when an order has expired or was cancelled, but the customer still wants to pay for it, or
if an order was reassigned, or otherwise modified and needs to be re-processed.

Arguments
Arguments for this route are set on the path, i.e.
`/reset_order/{order_id}`

Returns
The endpoint returns the order states before and after the reset.