feat: attach order
Attach an order to a Tari address

This endpoint is used to attach an order to a Tari address. This is useful when a user has made an order but
messed up the memo signature. Once support has verified the order, they can override the entry and attach
the order to the correct address.

The address does _not_ have to exist in the database. A new account will be created if the address is not found.

_Note_: This feature is written, but it's not clear how it should be supported. Currently, you _must_ supply a valid
memo signature with the order to prevent spoofing attacks. That signature contains the wallet address.
From there, you can use `reassign_order` to move the order to a different address.

So I'll leave this stub here for now, and once the server is in production, either a use case will present itself,
or we can remove this endpoint.