feat: exchange rates (#8)
Implemented ExchangeRateApi for managing exchange rates between Tari and other currencies.
The API provides functionality to fetch the last exchange rate for a given currency and to
set a new exchange rate.

The exchange rates are stored and retrieved using a database that implements the
`ExchangeRates` trait.

The Sqlite implementation is provided.

feat: exchange rate persistence

Adds a new `exchange_rates` table that stores the Tari <> Currency
exchange rate.

This is primarily intended for storefronts that don't allow custom
currencies, and you need to be able to convert from the base currency.

Two new routes are added as well, for retrieving and setting a rate in
the backend.

The propagation of this rate through to the storefront is not yet
implemented.

Updated base64

feat: exchange rate control

Created new crate, shopify_tools, to handle shopify-specific functions,
like querying an order, and setting the Tari<>USD exchange rate.

Added a new table, exchange rates, that cahces the current exchange rate
for Tari for various currencies. This table will be the source of truth
for currency conversions, and the server will take the responsibility of
keeping the shopify metaobject data (or other storefront equivalents)
up-to-date.

Some of this functionality is still stubbed out.

Adds new commands to taritools to perform order fetching, and exchange
rate management on shopify.