Claim tests (#11)
* test: claim order tests

Testing situation where an address is already associated with orders on
account A but then a claim is made with the address for an order on
account B.

Indicates need to allow many-to-many account:addresses

* feat: many-to-many accounts-addresses

Relaxes restriction that only one address can be associated per account.

This change allows the claim_order function to work, since the following
flow was previously disallowed:

- Make payment from address A (acc #1)
- Make order O without providing an wallet info with customer C (acc #2)
- Claim order O with address A (acc #2 != Acc #1. Error!)

Now we simply link A to acc #2 as well, and make the code aware of
future potential ambiguities.

fix: update tests

Orders with invalid memos still get placed now. They just get pt into
their own account.

This is better, since the orders will exost on Shopify, so they should
also exist here.