feat: update product hook
Adds the callback endpont for the `products/update` webhook.

This lets us check that the Tari price is correctly set whenever a new
product (not supported 100% yet) or variant (fully supported) is added.

Before setting the price metafield, the current value is checked against
the expected value.

When setting the Tari price rate, this hook is called for every variant
in the store, so the check is important to prevent an infinite loop.