Confirm payment (#24)
* feat: shopify event handlers

Adds shopify event handlers (OnOrderPaid, OnOrderAnnulled).

Adds a field to Orders table that captures the original order price in
the original currency. Generally only for informational purposes.

* feat: link shopify event handlers

Links the shopify event handlers to the main server instance.
In future, you'll want to make this conditional based on a configuration
file.

* Original price field

Adds an original_price field to the orders table. This is primarily for informational purposes (and is stored as a string).
For shopify it's used to send the expected amount field to the shopify API when marking an order as paid.

A few queries and tests were updated to accommodate this change.