feat: value parser
The miner notification passes a value like "35.005 T" instead of
MicroTari. So we add a parser to handle that.

Also fixes a bug in memo extraction which was causing the wrong log
message to be printed when a memo was present, but wasn't a
memoSignature

Also improves the tps_notify script to log more info for diagnostics.