tests: test order - claim - pay permutations
Adds tests that catch a bug where some db transactions were not being
committed.

fix: check all accounts on confirmation

When a payment is confirmed, we check all accounts associated with the
sender address to see if there are any orders that can be paid.

This required adding another API call,
`fetch_payable_oders_for_address`, which retrieves this list of orders.

Updated the definition of payable to "New" orders only. The tests showed
up an issue where "Unclaimed" orders would be paid erroneously under
certain conditions.

This change also uncovered a bug in the tests (a test was passing when
it shouldn't have because the claim signature was wrong. The test
correctly failed after this fix, and passed again wonce the signature
was updated to the new address format).

Added `fetch_payment_by_txid` method to the API. This is a fairly
fundamental call that was missing until now.