feat: refactor account management
The fact that "pay and claim" looks to become the norm rather than the
exception, I decided to rework how addresses and customers are linked.

Instead of an accounts table with two many-to-many connections to the
payments and orders tables respectively, the datbaase is changed to:

1. A single many-to-many association between payments and orders,
   faciliated via the address_customer_id_link table.
   The model here is that CUSTOMER IDs and ADDRESSes are linked, not
   orders and addresses.
   This significantly simplifies the logic in multiple places in the
   code and handles the claim and pay apporach much more robustly.
2. When payments are used to fulfil orders, these transactions are
   recorded in a `settlement_journal` table.

"Account" balances are now calculated via the `address_pending_balance`
and `address_balance` VIEWs, negating the need to manually manage the
accounting in the API, which was somewhat error-prone.

--

This PR also completed many TODOs:

* You can specify status(es) in the `/order_search` endpoint now, e.g.
  `/api/search/orders?status=Cancelled,Paid`
* Most, if not all, the TODOs in the cucumber tests have been completed.

* Updated taritools to handle the new endpoints and result types
* Updated 150+ tests for same