feat: account settlements
There may be edge cases where the order matching algorithms miss a
payment/order match. In this instance, we need a method to manually
retrigger the matching process.

The `settle` endpoints provide this feature.

Note that users can request a self-trigger for their connected wallets,
but we only allow admins to trigger settlements for general customer ids
and addresses to prevent a DoS vector.