feat: unsafe memos
This feature adds a new configuration variable,
TPG_DISABLE_MEMO_SIGNATURE_CHECK, which if enabled via
`TPG_DISABLE_MEMO_SIGNATURE_CHECK=1` will not require the memo field for
interactive payments to have be a MemoSignature attesting to the order
id.

In general, setting this flag may lead to spoofing attacks and is
discouraged, but it is added as a convenience for wallets that do not
want to add a call to create a signed memo field.