feat: alternative order id
Adds `alt-id` to the orders table, which provides an alternative means
of identifying orders. For example, Shopify uses the `order.id` field,
but presents `order.name` to clients.

Any reference to orders in the server can be configured to search both
the order_id and the alt_id fields by setting `TPG_STRICT_MODE=0` in the
configuration.